feat(storybook): migrate Switch story as pilot
